ClojureScript: handle inheritance correctly
Handle the case where an ornament component is used as the tagname of another
component, in ClojureScript. We need to find the parent component in our
registry, for which we need to expand the symbol to be fully qualified. In
Clojure we can do this via inspection of namespaces and vars, in ClojureScript
we get `(:ns &env)` and can work with that.

(o/defstyle foo x/bar ,,,)